+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Fixed
+
 - Fixed request crash with empty body and unexpected header Content-Type (#189)
+
+### Added
+
+- `roles.httpd` role to configure one or more HTTP servers (#196)
 
 ## [1.5.0] - 2023-03-29